Devlop -> main

See merge request s10-webmobile2-sub2/S10P12A210!76

+++ Back-End/src/main/java/com/backend/prog/domain/project/application/ProjectTechCodeService.java
@@ -1,9 +1,13 @@
 package com.backend.prog.domain.project.application;
 
+import com.backend.prog.domain.project.dao.ProjectMemberRespository;
 import com.backend.prog.domain.project.dao.ProjectTechCodeRespository;
-import com.backend.prog.domain.project.domain.Project;
 import com.backend.prog.domain.project.domain.ProjectCodeDetaliId;
+import com.backend.prog.domain.project.domain.ProjectMember;
+import com.backend.prog.domain.project.domain.ProjectMemberId;
 import com.backend.prog.domain.project.domain.ProjectTechCode;
+import com.backend.prog.global.error.CommonException;
+import com.backend.prog.global.error.ExceptionEnum;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -17,6 +21,7 @@ import java.util.Optional;
 public class ProjectTechCodeService {
 
     private final ProjectTechCodeRespository projectTechCodeRespository;
+    private final ProjectMemberRespository projectMemberRespository;
 
     public void createProjectTechCodeList(List<ProjectTechCode> list){
         projectTechCodeRespository.saveAll(list);
@@ -30,8 +35,18 @@ public class ProjectTechCodeService {
         projectTechCodeRespository.save(projectTechCode);
     }
 
-    public void deleteProject(Long projectId, Integer techCode) {
+    public void deleteProject(Long projectId, Integer memberId, Integer techCode) {
         ProjectCodeDetaliId projectCodeDetaliId = new ProjectCodeDetaliId(projectId, techCode);
+        
+        //요청한 멤버가 팀장인지 확인
+        ProjectMemberId projectMemberId = new ProjectMemberId(projectId, memberId);
+
+        Optional<ProjectMember> projectMember = projectMemberRespository.findById(projectMemberId);
+
+        if(!projectMember.isPresent() || projectMember.get().getRoleCode().getId() != 17){
+            throw new CommonException(ExceptionEnum.AUTHORITY_NOT_HAVE);
+        }
+
         projectTechCodeRespository.deleteById(projectCodeDetaliId);
     }
 }
